converted backend to TypeScript
